The rdborrow package

rdborrow is a package that implements causal inference methods for analyzing and designing clinical trials with external controls.

Authors

The following authors contribute to the development and maintainance of the package:

- Lei Shi, University of California Berkeley, leishi1998@gmail.com

- Herbert Pang, Genentech Inc. pathwayrf@gmail.com

- Chen Chen, Genentech Inc.

- Jiawen Zhu, Genentech Inc.

See Also

Useful links:

- GitHub Repo for rdborrow: https://github.com/pathwayrf/rdborrow

- Estimating treatment effect in randomized trial after control to treatment crossover using external controls: https://www.tandfonline.com/doi/full/10.1080/10543406.2024.2330209

- Causal estimators for incorporating external controls in randomized trials with longitudinal outcomes. In submission.